Add dynamic memory management (PMM, VMM, Heap) and SMP support across riscv64, riscv32, and arm64
- Implemented dynamic Physical Memory Manager (PMM) in kernel/pmm.c.
- Implemented buddy-like Kernel Heap allocator in kernel/heap.c.
- Implemented Spinlock synchronization primitives in kernel/spinlock.c.
- Implemented multicore CPU management in kernel/cpu.c.
- Implemented Virtual Memory Manager (VMM) supporting page table walks:
  - riscv64: Sv39 3-level paging with SMP support.
  - riscv32: Sv32 2-level paging.
  - arm64: AArch64 3-level paging with page descriptor type fixes.
- Refactored task.c on all architectures to use dynamic allocations instead of static tables for thread stack, code segments, and heaps.
- Abstracted architecture-specific console writes via a generic kputs helper.
- Cleaned up user ulib string.h and unistd.h headers to fix size_t conflicts on 32-bit.
- Fixed division warnings in gui.c on 32-bit targets.
- Updated Bazel rules to link spinlock, pmm, heap, and vmm across all targets.

+++ user/ulib/gui.c
@@ -6,11 +6,11 @@
 #include "stdio.h"
 
 static GLfixed pixel_to_ndc_x(int px, int width) {
-  return (GLfixed)((((int64_t)(px * 2 - width)) << 16) / width);
+  return (GLfixed)(((px * 2 - width) << 16) / width);
 }
 
 static GLfixed pixel_to_ndc_y(int py, int height) {
-  return (GLfixed)((((int64_t)(height - py * 2)) << 16) / height);
+  return (GLfixed)(((height - py * 2) << 16) / height);
 }
 
 void draw_rect_px(int x, int y, int w, int h, GLubyte r, GLubyte g, GLubyte b) {
